pyproject.toml
[project]
name = "cook-recipe"
version = "0.1.0"
description = ""
authors = [
    {name = "ReaperSans28",email = "smileuwu321@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-tinymce (>=5.0.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
